requirements.txt
Django==4.2.7
djangorestframework==3.14.0
Pillow==10.1.0
python-decouple==3.8
django-crispy-forms==2.1
crispy-bootstrap5==0.7

